+++ src/lib/components/dashboard/navbar/NavBar.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-	import { Cursor, Rectangle, Eraser, Arrow } from './navbar-icons';
+	import { Cursor, Rectangle, Eraser, Arrow, HandGrab } from './navbar-icons';
 	import { navBarState, keyPress, responsiveType } from '$lib/io/Stores';
 
-	let activeIndex: number | null = null;
+	let activeIndex: number | null | undefined = null;
 	let showTooltip: { [key: number]: boolean } = {};
 	let hoveredIndex: number | null = null;
 	let hoverTimeout: NodeJS.Timeout;
@@ -11,6 +11,7 @@
 	let clickedIndex: number | null = null;
 
 	let icons: { name: string; mode: NavBar; component: any; index: number; tooltip: string }[] = [
+		{ name: 'HandGrab', mode: 'pan', component: HandGrab, index: 0, tooltip: 'Pan Canvas' },
 		{
 			name: 'Cursor',
 			mode: 'select',
@@ -36,7 +37,7 @@
 			name: 'Eraser',
 			mode: 'eraser',
 			component: Eraser,
-			index: 0,
+			index: 4,
 			tooltip: 'Erase Canvas'
 		}
 	];
@@ -101,8 +102,44 @@
 	};
 </script>
 
-<div class="flex justify-center h-10 items-center rounded-md shadow-lg bg-neutral-800">
-	<div class="divide-x divide-neutral-700 flex items-center justify-items-center space-x-2">
+<div
+	class="flex justify-center items-center rounded-md shadow-lg h-12 bg-neutral-800 p-1 border border-1 border-neutral-700/70"
+>
+	<div class="divide-neutral-700 flex items-center justify-between space-x-2">
+		<div
+			on:click={() => setMode('pan', 0)}
+			on:mouseover={() => startHover(0)}
+			on:mouseout={endHover}
+			on:touchstart={() => startLongPress(0)}
+			on:touchend={endLongPress}
+			on:keypress={null}
+			on:blur={null}
+			on:focus={null}
+			class={`flex items-center w-8 h-8 justify-center mx-1 rounded-md overflow-hidden ${
+				activeIndex === 0 ? 'fill-blue-500 ' : ''
+			} ${
+				$responsiveType !== 'touch' || isPressed
+					? 'hover:text-neutral-700/90 hover:bg-neutral-500'
+					: ''
+			}`}
+		>
+			<HandGrab
+				class="w-5 h-5 ml-1  text-neutral-400 hover:text-neutral-200 active:text-neutral-50"
+			/>
+			<div class="bottom-0 right-0 mt-4">
+				<span class="text-xs text-neutral-500">0</span>
+			</div>
+
+			{#if showTooltip[0] && (clickedIndex !== 0 || $responsiveType !== 'touch')}
+				<div
+					role="tooltip"
+					class="absolute -bottom-6 left-1/2 z-30 transform -translate-x-1/2 px-1 bg-neutral-200 text-gray-700 text-xs shadow-sm"
+				>
+					Select Items
+				</div>
+			{/if}
+		</div>
+
 		<div
 			on:click={() => setMode('select', 1)}
 			on:mouseover={() => startHover(1)}
@@ -112,7 +149,7 @@
 			on:keypress={null}
 			on:blur={null}
 			on:focus={null}
-			class={`flex items-center  justify-center mx-1 rounded-md overflow-hidden ${
+			class={`flex items-center w-8 h-8 justify-center mx-1 rounded-md overflow-hidden ${
 				activeIndex === 1
 					? 'bg-[#6f6599]'
 					: $responsiveType !== 'touch' || isPressed
@@ -120,7 +157,13 @@
 					: ''
 			}`}
 		>
-			<Cursor />
+			<Cursor
+				class="w-5 ml-1  h-5 text-neutral-400 hover:text-neutral-200 active:text-neutral-50"
+			/>
+			<div class="bottom-0 right-0 mt-4">
+				<span class="text-xs text-neutral-500">1</span>
+			</div>
+
 			{#if showTooltip[1] && (clickedIndex !== 1 || $responsiveType !== 'touch')}
 				<div
 					role="tooltip"
@@ -131,7 +174,7 @@
 			{/if}
 		</div>
 
-		<div class="flex items-center justify-items-center">
+		<div class="flex items-center justify-items-center space-x-3">
 			<div
 				on:click={() => setMode('drawRectangle', 2)}
 				on:keypress={null}
@@ -141,15 +184,21 @@
 				on:touchend={endLongPress}
 				on:blur={null}
 				on:focus={null}
-				class={`flex items-center  justify-center mx-1 rounded-md overflow-hidden ${
+				class={`flex items-center w-8 h-8 justify-center mx-1 rounded-md overflow-hidden ${
 					activeIndex === 2
-						? 'bg-[#6f6599]'
+						? 'bg-[#6f6599] fill-blue-400'
 						: $responsiveType !== 'touch' || isPressed
 						? 'hover:text-neutral-700/90 hover:bg-neutral-500'
 						: ''
 				}`}
 			>
-				<Rectangle />
+				<Rectangle
+					class="w-5 ml-1 h-5 text-neutral-400 hover:text-neutral-200 active:text-neutral-50 "
+				/>
+				<div class="bottom-0 right-0 mt-4">
+					<span class="text-xs text-neutral-500">2</span>
+				</div>
+
 				{#if showTooltip[2]}
 					<div
 						role="tooltip"
@@ -170,7 +219,7 @@
 				on:touchend={endLongPress}
 				on:blur={null}
 				on:focus={null}
-				class={`flex items-center  justify-center mx-1 rounded-md overflow-hidden ${
+				class={`flex items-center w-8 h-8 justify-center mx-1 rounded-md overflow-hidden ${
 					activeIndex === 3
 						? 'bg-[#6f6599]'
 						: $responsiveType !== 'touch' || isPressed
@@ -178,7 +227,13 @@
 						: ''
 				}`}
 			>
-				<Arrow />
+				<Arrow
+					class="w-5 h-5 ml-1 fill:red-400 text-neutral-400 hover:text-neutral-200 active:text-neutral-50"
+				/>
+				<div class="bottom-0 right-0 mt-4">
+					<span class="text-xs text-neutral-500">3</span>
+				</div>
+
 				{#if showTooltip[3]}
 					<div
 						role="tooltip"
@@ -191,24 +246,29 @@
 
 			<!-- Eraser Icon -->
 			<div
-				on:click={() => setMode('eraser', 0)}
+				on:click={() => setMode('eraser', 4)}
 				on:keypress={null}
-				on:mouseover={() => startHover(0)}
+				on:mouseover={() => startHover(4)}
 				on:mouseout={endHover}
-				on:touchstart={() => startLongPress(0)}
+				on:touchstart={() => startLongPress(4)}
 				on:touchend={endLongPress}
 				on:blur={null}
 				on:focus={null}
-				class={`flex items-center  justify-center mx-1 rounded-md overflow-hidden ${
-					activeIndex === 0
+				class={`flex items-center w-8 h-8 justify-center mx-1 rounded-md overflow-hidden ${
+					activeIndex === 4
 						? 'bg-[#6f6599]'
 						: $responsiveType !== 'touch' || isPressed
 						? 'hover:text-neutral-700/90 hover:bg-neutral-500'
 						: ''
 				}`}
 			>
-				<Eraser />
-				{#if showTooltip[0]}
+				<Eraser
+					class="w-5 h-5 ml-1 text-neutral-400 hover:text-neutral-200 active:text-neutral-50"
+				/>
+				<div class="bottom-0 right-0 mt-4">
+					<span class="text-xs text-neutral-500">4</span>
+				</div>
+				{#if showTooltip[4]}
 					<div
 						role="tooltip"
 						class="absolute -bottom-6 left-1/2 z-30 transform -translate-x-1/2 px-1 bg-neutral-200 text-gray-700 text-xs shadow-sm"