+++ src/lib/components/dashboard/navbar/navbar-icons/index.ts
@@ -3,5 +3,6 @@ import Rectangle from '../../../ui/icons/Rectangle.svelte';
 import Eraser from '../../../ui/icons/Eraser.svelte';
 import Text from '../../../ui/icons/Text.svelte';
 import Arrow from '../../../ui/icons/Arrow.svelte';
+import HandGrab from '../../../ui/icons/HandGrab.svelte';
 
-export { Cursor, Rectangle, Eraser, Text, Arrow };
+export { Cursor, Rectangle, Eraser, Text, Arrow, HandGrab };