+++ ReleaseNotes.md
@@ -2,13 +2,14 @@
 
 ## October 2023
 
+- Added Zooming and Infinite Canvas
 - Disables BUttons without a chart.
 - Created new landing page.
 - Optimized the queries so it only runs once.
 - Added UI/UX layering
 - fixed dropzone file drag over
 - made the mobile smoother.
-- Fixed INDEXDB annoying bug.
+- Fixed the INDEXDB annoying bug.
 - Can persist file handles with DuckDB
 
 ## September 2023